// src/api/schema.ts
export type LocalizedString = Record<string, string>;

export type MangaResponse = {
    result: 'ok' | 'error'
    /** Default: "entity" */
    response: string
    data: Manga
};

export type ChapterResponse = {
    result: 'ok' | 'error'
    /** Default: "entity" */
    response: string
    data: Chapter
};

export type Relationship = {
    /** UUID formatted string */
    id: string
    type: string
    /** Only present if you are on a Manga entity and a Manga relationship */
    related: 'monochrome' | 'main_story' | 'adapted_from' | 'based_on' | 'prequel' | 'side_story' | 'doujinshi' | 'same_franchise' | 'shared_universe' | 'sequel' | 'spin_off' | 'alternate_story' | 'alternate_version' | 'preserialization' | 'colored' | 'serialization'
    /** If Reference Expansion is applied, contains objects attributes */
    attributes: any | null
};

export type Chapter = {
    /** UUID formatted string */
    id: string
    type: 'chapter'
    attributes: ChapterAttributes
    relationships: Relationship[]
};

export type Manga = {
    /** UUID formatted string */
    id: string
    type: 'manga'
    attributes: MangaAttributes
    relationships: Relationship[]
};
// ...
export type MangaAttributes = {
    title: LocalizedString
    altTitles: LocalizedString[]
    description: LocalizedString
    isLocked: boolean
    links: Links
    originalLanguage: string
    lastVolume: string | null
    lastChapter: string | null
    publicationDemographic: 'shounen' | 'shoujo' | 'josei' | 'seinen' | null
    status: 'completed' | 'ongoing' | 'cancelled' | 'hiatus'
    year: number | null
    contentRating: 'safe' | 'suggestive' | 'erotica' | 'pornographic'
    chapterNumbersResetOnNewVolume: boolean
    availableTranslatedLanguages: any[]
    /** UUID formatted string */
    latestUploadedChapter: string
    tags: Tag[]
    state: 'draft' | 'submitted' | 'published' | 'rejected'
    /** ```console
     * Minimum: 1
     * ``` */
    version: number
    createdAt: string
    updatedAt: string
};
// ...
export type Tag = {
    /** UUID formatted string */
    id: string
    type: 'tag'
    attributes: TagAttributes
    relationships: Relationship[]
};